Cleaning up, again
Renegade dev 1.1.84
Bench: 3330333

+++ Renegade/Reporting.h
@@ -8,16 +8,16 @@ struct Results {
 	int depth = 0;
 	int seldepth = 0;
 	uint64_t nodes = 0;
-	int time = 0;
-	int nps = 0;
+	uint64_t time = 0;
+	uint64_t nps = 0;
 	int hashfull = 0;
 	int ply = 0;
 	std::vector<Move> pv;
 	int threads = 1;
 
 	Results();
-	Results(const int score, const int depth, const int seldepth, const uint64_t nodes, const int time,
-		const int nps, const int hashfull, const int ply, const std::vector<Move>& pv);
+	Results(const int score, const int depth, const int seldepth, const uint64_t nodes, const uint64_t time,
+		const uint64_t nps, const int hashfull, const int ply, const std::vector<Move>& pv);
 	Move BestMove() const;
 };
